refactor(mfa): move the device check fully out of the biometrics hooks
The validateDevice actor owns the device check now, so the hooks no
longer need to expose deviceVerificationType,
doesDeviceSupportAuthenticationMethod, or deviceCheckFailureReason.
useNavigateTo3DSAuthorizationChallenge, their last consumer, reads the
two values it needs straight from the platform-resolved operations
module, which also drops its useBiometrics call and the as cast on the
allowed-methods list. The device-support tests move from the HSM hook
test to a dedicated operations test.
