+++ src/components/MultifactorAuthentication/biometrics/operations/index.native.ts
@@ -5,10 +5,10 @@ import CONST from '@src/CONST';
 import {isSensorAvailable} from '@sbaiahmed1/react-native-biometrics';
 
 /**
- * Platform-resolved biometric operations shared by the MFA machine actors and the React biometrics
- * hooks. These functions read no Onyx and no React state, so a machine actor can import them directly
- * instead of reaching through a hook. Native resolves to this file and web resolves to the `index.ts`
- * sibling. Later slices add the registration and authorization operations here.
+ * Platform-resolved biometric operations for the device check. These functions read no Onyx and no
+ * React state, so the MFA machine actors and other non-React callers can import them directly. Native
+ * resolves to this file and web resolves to the `index.ts` sibling. Later slices add the registration
+ * and authorization operations here.
  */
 
 /** The authentication method this platform verifies with. Native verifies with HSM-backed biometrics. */
